[WIP] Reworked architecture to use a manager
