+++ .env.example
@@ -1,8 +1,8 @@
 # Copy this file to .env and edit with your own values
 
 # SpigotMC Credentials
-LOGIN = 'example@example.com'
-PASSWORD = 'password'
+SPIGOTMC_LOGIN = 'example@example.com'
+SPIGOTMC_PASSWORD = 'password'
 
 CLOUDPROXY_URL = 'http://localhost:8191/v1'
