+++ download/manager.py
@@ -1,0 +1,9 @@
+
+from download.sources.spigotmc_source import SpigotMcSource
+
+
+class DownloadManager():
+    spigotmc_source = None
+
+    def __init__(self):
+        self.spigotmc_source = SpigotMcSource()
