+++ download/sources/base_source.py
@@ -1,0 +1,6 @@
+import requests
+
+class BaseSource:
+
+    def download_element(self, url):
+        return requests.get(url).content
