+++ requirements.txt
@@ -1,3 +1,4 @@
+pylint==2.6.0
 python-dotenv==0.14.0
 beautifulsoup4==4.9.1
 PyYAML==5.3.1
