+++ setup.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+rm -rf env
 python3 -m venv env
 source env/bin/activate
 pip install -r requirements.txt
